Moved models and notebooks to a single directory
